Adopt `justfile` for CI; run `cargo clippy --fix` (#139)
* Use Just for linting

* Run `cargo clippy --fix`

* Update Justfile

* Update CI to use justfile


* Pre-install binstall

* Modify justfile for CI

* Fix it

* Install nightly rustfmt


* Install nightly rustfmt


* Install nightly rustfmt


* Try to fix test step


* Reorder stuff


---------

+++ crates/fefix/src/tagvalue/encoder.rs
@@ -157,7 +157,7 @@ where
     where
         V: FieldType<'s>,
     {
-        self.set_with(tag.get() as u32, value, settings)
+        self.set_with(tag.get(), value, settings)
     }
 }
